+++ pkg/handlers/aws/mutation/instancetype/inject_control_plane_test.go
@@ -1,11 +1,9 @@
 // Copyright 2023 D2iQ, Inc. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package region
+package instancetype
 
 import (
-	"testing"
-
 	. "github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
 	runtimehooksv1 "sigs.k8s.io/cluster-api/exp/runtime/hooks/api/v1alpha1"
@@ -17,15 +15,9 @@ import (
 	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/pkg/handlers/generic/clusterconfig"
 )
 
-func TestRegionPatch(t *testing.T) {
-	gomega.RegisterFailHandler(Fail)
-	RunSpecs(t, "AWS Region mutator suite")
-}
-
-var _ = Describe("Generate AWS Region patches", func() {
-	// only add aws region patch
+var _ = Describe("Generate InstanceType patches for ControlPlane", func() {
 	patchGenerator := func() mutation.GeneratePatches {
-		return mutation.NewMetaGeneratePatchesHandler("", NewPatch()).(mutation.GeneratePatches)
+		return mutation.NewMetaGeneratePatchesHandler("", NewControlPlanePatch()).(mutation.GeneratePatches)
 	}
 
 	testDefs := []capitest.PatchTestDef{
@@ -33,20 +25,21 @@ var _ = Describe("Generate AWS Region patches", func() {
 			Name: "unset variable",
 		},
 		{
-			Name: "region set",
+			Name: "instanceType for controlplane set",
 			Vars: []runtimehooksv1.Variable{
 				capitest.VariableWithValue(
 					clusterconfig.MetaVariableName,
-					"a-specific-region",
+					"m5.xlarge",
+					clusterconfig.MetaControlPlaneConfigName,
 					v1alpha1.AWSVariableName,
 					VariableName,
 				),
 			},
-			RequestItem: request.NewAWSClusterTemplateRequestItem("1234"),
+			RequestItem: request.NewCPAWSMachineTemplateRequestItem("1234"),
 			ExpectedPatchMatchers: []capitest.JSONPatchMatcher{{
-				Operation:    "add",
-				Path:         "/spec/template/spec/region",
-				ValueMatcher: gomega.Equal("a-specific-region"),
+				Operation:    "replace",
+				Path:         "/spec/template/spec/instanceType",
+				ValueMatcher: gomega.Equal("m5.xlarge"),
 			}},
 		},
 	}
