+++ pkg/handlers/aws/mutation/instancetype/inject_worker_test.go
@@ -1,31 +1,24 @@
 // Copyright 2023 D2iQ, Inc. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package region
+package instancetype
 
 import (
-	"testing"
-
 	. "github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
+	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	runtimehooksv1 "sigs.k8s.io/cluster-api/exp/runtime/hooks/api/v1alpha1"
 
 	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/api/v1alpha1"
 	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/common/pkg/capi/clustertopology/handlers/mutation"
 	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/common/pkg/testutils/capitest"
 	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/common/pkg/testutils/capitest/request"
-	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/pkg/handlers/generic/clusterconfig"
+	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/pkg/handlers/generic/workerconfig"
 )
 
-func TestRegionPatch(t *testing.T) {
-	gomega.RegisterFailHandler(Fail)
-	RunSpecs(t, "AWS Region mutator suite")
-}
-
-var _ = Describe("Generate AWS Region patches", func() {
-	// only add aws region patch
+var _ = Describe("Generate InstanceType patches for Worker", func() {
 	patchGenerator := func() mutation.GeneratePatches {
-		return mutation.NewMetaGeneratePatchesHandler("", NewPatch()).(mutation.GeneratePatches)
+		return mutation.NewMetaGeneratePatchesHandler("", NewWorkerPatch()).(mutation.GeneratePatches)
 	}
 
 	testDefs := []capitest.PatchTestDef{
@@ -33,20 +26,26 @@ var _ = Describe("Generate AWS Region patches", func() {
 			Name: "unset variable",
 		},
 		{
-			Name: "region set",
+			Name: "instanceType for workers set",
 			Vars: []runtimehooksv1.Variable{
 				capitest.VariableWithValue(
-					clusterconfig.MetaVariableName,
-					"a-specific-region",
+					workerconfig.MetaVariableName,
+					"m5.2xlarge",
 					v1alpha1.AWSVariableName,
 					VariableName,
 				),
+				capitest.VariableWithValue(
+					"builtin",
+					apiextensionsv1.JSON{
+						Raw: []byte(`{"machineDeployment": {"class": "a-worker"}}`),
+					},
+				),
 			},
-			RequestItem: request.NewAWSClusterTemplateRequestItem("1234"),
+			RequestItem: request.NewWorkerAWSMachineTemplateRequestItem("1234"),
 			ExpectedPatchMatchers: []capitest.JSONPatchMatcher{{
-				Operation:    "add",
-				Path:         "/spec/template/spec/region",
-				ValueMatcher: gomega.Equal("a-specific-region"),
+				Operation:    "replace",
+				Path:         "/spec/template/spec/instanceType",
+				ValueMatcher: gomega.Equal("m5.2xlarge"),
 			}},
 		},
 	}
