+++ pkg/handlers/aws/mutation/securitygroups/inject_control_plane_test.go
@@ -1,13 +1,12 @@
 // Copyright 2023 D2iQ, Inc. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package region
+package securitygroups
 
 import (
-	"testing"
-
 	. "github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
+	"k8s.io/utils/ptr"
 	runtimehooksv1 "sigs.k8s.io/cluster-api/exp/runtime/hooks/api/v1alpha1"
 
 	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/api/v1alpha1"
@@ -17,15 +16,9 @@ import (
 	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/pkg/handlers/generic/clusterconfig"
 )
 
-func TestRegionPatch(t *testing.T) {
-	gomega.RegisterFailHandler(Fail)
-	RunSpecs(t, "AWS Region mutator suite")
-}
-
-var _ = Describe("Generate AWS Region patches", func() {
-	// only add aws region patch
+var _ = Describe("Generate SecurityGroup patches for ControlPlane", func() {
 	patchGenerator := func() mutation.GeneratePatches {
-		return mutation.NewMetaGeneratePatchesHandler("", NewPatch()).(mutation.GeneratePatches)
+		return mutation.NewMetaGeneratePatchesHandler("", NewControlPlanePatch()).(mutation.GeneratePatches)
 	}
 
 	testDefs := []capitest.PatchTestDef{
@@ -33,21 +26,38 @@ var _ = Describe("Generate AWS Region patches", func() {
 			Name: "unset variable",
 		},
 		{
-			Name: "region set",
+			Name: "SecurityGroups for controlplane set",
 			Vars: []runtimehooksv1.Variable{
 				capitest.VariableWithValue(
 					clusterconfig.MetaVariableName,
-					"a-specific-region",
+					v1alpha1.AdditionalSecurityGroup{
+						{ID: ptr.To("sg-1")},
+						{ID: ptr.To("sg-2")},
+						{ID: ptr.To("sg-3")},
+					},
+					clusterconfig.MetaControlPlaneConfigName,
 					v1alpha1.AWSVariableName,
 					VariableName,
 				),
 			},
-			RequestItem: request.NewAWSClusterTemplateRequestItem("1234"),
-			ExpectedPatchMatchers: []capitest.JSONPatchMatcher{{
-				Operation:    "add",
-				Path:         "/spec/template/spec/region",
-				ValueMatcher: gomega.Equal("a-specific-region"),
-			}},
+			RequestItem: request.NewCPAWSMachineTemplateRequestItem("1234"),
+			ExpectedPatchMatchers: []capitest.JSONPatchMatcher{
+				{
+					Operation:    "add",
+					Path:         "/spec/template/spec/additionalSecurityGroups",
+					ValueMatcher: gomega.HaveLen(3),
+				},
+				// TODO(shalinpatel): add matcher to check if all SG are set
+				// {
+				// 	Operation: "add",
+				// 	Path:      "/spec/template/spec/additionalSecurityGroups",
+				// 	ValueMatcher: gomega.ContainElements(
+				// 		gomega.HaveKeyWithValue("id", "sg-1"),
+				// 		gomega.HaveKeyWithValue("id", "sg-2"),
+				// 		gomega.HaveKeyWithValue("id", "sg-3"),
+				// 	),
+				// },
+			},
 		},
 	}
