--- pkg/handlers/generic/mutation/users/tests/generate_patches.go
@@ -1,97 +1,0 @@
-// Copyright 2023 D2iQ, Inc. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package tests
-
-import (
-	"testing"
-
-	"github.com/onsi/gomega"
-	"k8s.io/apiserver/pkg/storage/names"
-	runtimehooksv1 "sigs.k8s.io/cluster-api/exp/runtime/hooks/api/v1alpha1"
-
-	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/api/v1alpha1"
-	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/common/pkg/capi/clustertopology/handlers/mutation"
-	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/common/pkg/testutils/capitest"
-	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/common/pkg/testutils/capitest/request"
-)
-
-var (
-	testUser1 = v1alpha1.User{
-		Name:           "complete",
-		HashedPassword: "password",
-		SSHAuthorizedKeys: []string{
-			"key1",
-			"key2",
-		},
-		Sudo: "ALL=(ALL) NOPASSWD:ALL",
-	}
-	testUser2 = v1alpha1.User{
-		Name: "onlyname",
-	}
-)
-
-func TestGeneratePatches(
-	t *testing.T,
-	generatorFunc func() mutation.GeneratePatches,
-	variableName string,
-	variablePath ...string,
-) {
-	t.Helper()
-
-	capitest.ValidateGeneratePatches(
-		t,
-		generatorFunc,
-		capitest.PatchTestDef{
-			Name: "unset variable",
-		},
-		capitest.PatchTestDef{
-			Name: "users set for KubeadmControlPlaneTemplate",
-			Vars: []runtimehooksv1.Variable{
-				capitest.VariableWithValue(
-					variableName,
-					[]v1alpha1.User{testUser1, testUser2},
-					variablePath...,
-				),
-			},
-			RequestItem: request.NewKubeadmControlPlaneTemplateRequestItem(""),
-			ExpectedPatchMatchers: []capitest.JSONPatchMatcher{{
-				Operation:    "add",
-				Path:         "/spec/template/spec/kubeadmConfigSpec/users",
-				ValueMatcher: gomega.HaveLen(2),
-			}},
-		},
-	)
-
-	capitest.ValidateGeneratePatches(
-		t,
-		generatorFunc,
-		capitest.PatchTestDef{
-			Name: "unset variable",
-		},
-		capitest.PatchTestDef{
-			Name: "users set for KubeadmConfigTemplate generic worker",
-			Vars: []runtimehooksv1.Variable{
-				capitest.VariableWithValue(
-					variableName,
-					[]v1alpha1.User{testUser1, testUser2},
-					variablePath...,
-				),
-				capitest.VariableWithValue(
-					"builtin",
-					map[string]any{
-						"machineDeployment": map[string]any{
-							"class": names.SimpleNameGenerator.GenerateName("worker-"),
-						},
-					},
-				),
-			},
-			RequestItem: request.NewKubeadmConfigTemplateRequestItem(""),
-			ExpectedPatchMatchers: []capitest.JSONPatchMatcher{{
-				Operation:    "add",
-				Path:         "/spec/template/spec/users",
-				ValueMatcher: gomega.HaveLen(2),
-			}},
-		},
-	)
-}
