--- pkg/handlers/nutanix/mutation/controlplaneendpoint/tests/generate_patches.go
@@ -1,59 +1,0 @@
-// Copyright 2023 D2iQ, Inc. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package tests
-
-import (
-	"testing"
-
-	"github.com/onsi/gomega"
-	clusterv1 "sigs.k8s.io/cluster-api/api/v1beta1"
-	runtimehooksv1 "sigs.k8s.io/cluster-api/exp/runtime/hooks/api/v1alpha1"
-
-	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/common/pkg/capi/clustertopology/handlers/mutation"
-	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/common/pkg/testutils/capitest"
-	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/common/pkg/testutils/capitest/request"
-)
-
-func TestGeneratePatches(
-	t *testing.T,
-	generatorFunc func() mutation.GeneratePatches,
-	variableName string,
-	variablePath ...string,
-) {
-	t.Helper()
-
-	capitest.ValidateGeneratePatches(
-		t,
-		generatorFunc,
-		capitest.PatchTestDef{
-			Name: "unset variable",
-		},
-		capitest.PatchTestDef{
-			Name: "ControlPlaneEndpoint set to valid host and port",
-			Vars: []runtimehooksv1.Variable{
-				capitest.VariableWithValue(
-					variableName,
-					clusterv1.APIEndpoint{
-						Host: "10.20.100.10",
-						Port: 6443,
-					},
-					variablePath...,
-				),
-			},
-			RequestItem: request.NewNutanixClusterTemplateRequestItem(""),
-			ExpectedPatchMatchers: []capitest.JSONPatchMatcher{
-				{
-					Operation:    "replace",
-					Path:         "/spec/template/spec/controlPlaneEndpoint/host",
-					ValueMatcher: gomega.Equal("10.20.100.10"),
-				},
-				{
-					Operation:    "replace",
-					Path:         "/spec/template/spec/controlPlaneEndpoint/port",
-					ValueMatcher: gomega.BeEquivalentTo(6443),
-				},
-			},
-		},
-	)
-}
